+++ src/utils.ts
@@ -1,0 +1,72 @@
+import * as vscode from 'vscode';
+import { NeuroClient } from "neuro-game-sdk";
+
+import { NEURO } from './constants';
+
+export function assert(obj: any): asserts obj {
+    if(!obj)
+        throw new Error('Assertion failed');
+}
+
+export function logOutput(tag: string, message: string) {
+    if(!NEURO.outputChannel) {
+        console.error('Output channel not initialized');
+        return;
+    }
+    let ms = Date.now() % 1000;
+    let time = new Date().toLocaleTimeString('en-US', { hour12: false, hour: '2-digit', minute: '2-digit', second: '2-digit'}) + '.' + ms.toString().padStart(3, '0');
+    const prefix = `${time} [${tag}] `;
+    for(const line of message.split('\n')) {
+        NEURO.outputChannel.appendLine(prefix + line);
+    }
+}
+
+export function createClient() {
+    logOutput('INFO', 'Creating Neuro API client');
+    if(NEURO.client)
+        NEURO.client.disconnect();
+
+    NEURO.connected = false;
+    NEURO.waiting = false;
+    NEURO.cancelled = false;
+
+    // TODO: Check if this is a memory leak
+    NEURO.client = new NeuroClient(NEURO.url, NEURO.gameName, () => {
+        assert(NEURO.client instanceof NeuroClient);
+
+        logOutput('INFO', 'Connected to Neuro API');
+        NEURO.connected = true;
+
+        vscode.window.showInformationMessage('Successfully connected to Neuro API.');
+
+        NEURO.client.sendContext(
+            vscode.workspace.getConfiguration('neuropilot').get('initialContext', 'Something went wrong, blame whoever made this extension.'),
+        );
+
+        NEURO.client.onClose = () => {
+            NEURO.connected = false;
+            logOutput('INFO', 'Disconnected from Neuro API');
+            vscode.window.showInformationMessage('Disconnected from Neuro API.');
+        };
+
+        NEURO.client.onError = (error) => {
+            logOutput('ERROR', `Neuro client error: ${error}`);
+            vscode.window.showErrorMessage(`Neuro client error: ${error}`);
+        };
+
+        for(const handler of clientCreatedHandlers) {
+            handler();
+        }
+    });
+
+    NEURO.client.onError = () => {
+        logOutput('ERROR', 'Could not connect to Neuro API');
+        vscode.window.showErrorMessage('Could not connect to Neuro API.');
+    };
+}
+
+let clientCreatedHandlers: (() => void)[] = [];
+
+export function onClientCreated(handler: () => void) {
+    clientCreatedHandlers.push(handler);
+}
